Fix Timedelta tests
* fixtures cannot yield generators
* use a python built-in random number generator with constant seed
* split power-of-ten and random number tests for precise_string-roundtrip
* test precise_string-roundtrips for the zero timestamp
